[DEBUG] Keep latest source address as neighbor's address
In RFC5340, it is stated:
  "For all link types except virtual links, the Neighbor IP address is
   set to the IPv6 source address in the IPv6 header of the received OSPF
   Hello packet."

It is not said which source address should be kept among all received
Hello packets from the same neighbor. In current implementation, the first
is used.

In some cases, the neighbor's source address can change. But, in OSPFv3,
neighbors are always identified by their IDs. If we don't keep the latest
address as the neighbor's address, all HELLO packets will be correctly
considered, but further unicast packet sent to the router that changed its
address may not be received, making the link state blocking indefinitly.

Cases where the source address can change are:
- A wire's end is moved from one port to another.
- For some reason, a router decides to change its Active Interface, which
choice is implementation dependent.